ececise write on what to do is
what is the exerciseses

+++ Udemy-Java-Kurs-InputOutput/src/Auto.java
@@ -1,11 +1,11 @@
 
 public class Auto {
 
-	int reifen;  // ein Objekt mit dem new Operator erzeugen wir ein Objekt in dieser Klasse
-	int preis;
+	int reifen;            // ein Objekt mit dem new Operator erzeugen wir ein Objekt in dieser Klasse
+	double preis;
 	
 	
-	
+	// schauen es mit der Main methode an weil das ist eine seperate klasse wo als referenz zur main dient.
